voi_glmn function

voi_glmn function

voi_glm(
  n_sim,
  thresholds = (0:99)/100,
  glm_object = NULL,
  mc_type = "bootstrap",
  data = NULL,
  pi = NULL,
  truth_formula = NULL,
  family = binomial(link = "logit")
)

Arguments

n_sim

Number of simulations required.

thresholds

thresholds at which EVPI is calculated

glm_object

The GLM object representing the proposed model

mc_type

any of "bootstrap", "Bayesian_bootstrap", or "likelihood"

data

for EVPI calculations.

pi

A vector of predicted values. Required only if data is provided. If a string or single number it indicates the corresponding columns in data

truth_formula

Formula of the correct model. Its parameters will be estimated repeatedly using the data

family

GLM family and link function